API (REST) - Flat advert CRUD
API (REST) - Flat advert CRUD

+++ src/BuyHouse/BuyHouse.BLL/DTO/FlatAdvertFilter.cs
@@ -1,9 +1,4 @@
 ﻿using BuyHouse.DAL.Entities.HelperEnum;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace BuyHouse.BLL.DTO
 {
@@ -11,8 +6,8 @@ namespace BuyHouse.BLL.DTO
     {
         public string CityName { get; set; }
         public string CountRooms { get; set; }
-        public int MinPrice { get; set; }
-        public int MaxPrice { get; set; }
+        public ulong MinPrice { get; set; }
+        public ulong MaxPrice { get; set; }
         public Currency Currency { get; set; }
         public TypeOfPrice TypeOfPrice { get; set; }
         public double MinTotalArea { get; set; }
