+++ src/BuyHouse/BuyHouse.DAL/Entities/AdvertEntities/HouseAdvert.cs
@@ -28,8 +28,8 @@ namespace BuyHouse.DAL.Entities.AdvertEntities
         public string? CadastralNumber { get; set; }
 
         /*Advert properties*/
-        public int? TotalPrice { get; set; }
-        public int? PricePerSquareMeter { get; set; }
+        public ulong? TotalPrice { get; set; }
+        public ulong? PricePerSquareMeter { get; set; }
         public Currency Currency { get; set; }
         public TypeOfPrice TypePrice { get; set; }
         public DateTime? CreationDate { get; set; }
